Change test to reflect new goal for disposal behavior.
My reasoning for this goes: we don't know whether the immediate part we're disposing would need to dispose of the subcontainers before, during, or after its own disposal. So MEF shouldn't be doing it at all. The disposed part itself should have tracked the subcontainers and disposed it by its own policy. And in fact MEFv2 didn't do it at all, so we need a compelling reason to depart from established behavior.