Add ability to load MEF component assemblies by path.
VS MEF previously had this ability to load assemblies into the AppDomain based on the path in which they were found during the discovery phase. This adds the ability to MEFv3 as well.
Since AssemblyName.CodeBasePath is only available when targeting .NET 4.5, I put this ability in MS.VS.Composition.Configuration.dll that targets that framework and it's an extension method on ComposableCatalog that users can add. Otherwise, we default to the behavior of relying on the standard CLR probing paths to find the assembly.

I also take this opportunity to remove the generated GetAssemblyName method with its big switch statement in favor of just initializing an array in the generated code.