Runtime ExportFactory with lazy loading.
Retrofited the RuntimeExportProviderFactory to run on a new RuntimeComposition that describes the graph minimally and avoiding all uses of Type or other types that would force an assembly load before it is absolutely necessary.

At this point, the code compiles, but generates an assembly whose Reflection.TypeRef struct cannot be loaded.