Avoid loading assemblies for equality checks.
Although the newly added test did not fail prior to the product change, to engineer a test that would fail first would require a hash collision that would be very difficult to engineer in a test. I saw the failure in VS and the fix is straightforward by simply generalizing a fix we applied to only one place.