Avoid CLR COMDelegate::DelegateConstruct slow path.
ETL traces show this in these three methods that I change in this commit, and David Wrighton suggested the workaround is to create the delegate around an instance method rather than a static one (in other words, use the lambda syntax).

In the future, we hope to better understand the slow path and when it is chosen so that we can use the more efficient static method approach (and avoid the closure) whenever possible and fallback to the lambda when necessary to avoid the slow path.