Avoid unnecessary immutable and hashing overhead.
We want to avoid repeating the work done by enumeration, but we needn't use ImmutableHashSet to cache it. That incurs a hashing cost as well as the cost to allocate a bunch of objects for immutable data structures. We just need a simple collection.

This should shave off 14ms from AusGov solution open.