Remove string "compression" during serialization.
The StringBuilder we created on each invocation produced 1.2GB of garbage in construction alone. We could reuse instances, but I realize now that deserializing compressed strings loses the built in opportunity for string interning. So let's serialize the whole string, once, so that on deserialization each string is represented in the GC heap exactly once.