Fix leak in disposal of child ExportProviders.
ROOT CAUSE:
Sub-scope parts can import parts from their parent scope. When they lazily import a part from a parent scope that has not yet been created, the child ExportProvider would construct the value factory for that shared part and then store it in the parent ExportProvider's dictionary of shared instances. That valu factory had 'captured' a reference to the child ExportProvider and as a result, the child ExportProvider's lifetime was now tied to that of the longer lived parent.

FIX:
Avoid allowing "this" to be captured in the closure of the value factory.
Also avoid supplying provisionalSharedObjects for these value factories when they are passed "up" the ExportProvider hierarchy as well since those objects may be or refer to other short-lived child scoped objects.

As part of removing "this" from the closure, a Dev12 C# compiler bug still captured "this" even though I stopped using it. The Roslyn C# compiler has this bug fixed, but to make sure the code performs correctly regardless of compiler version I split the method in two to workaround the compiler bug.

This fix had a significant ramification: it stopped MEFv3's ability to magically allow the construction of shared parts at inappropriate sharing boundaries. This caused the InvalidImportAcrossSharingBoundaryV3 test to start failing, which documented how VS MEF performed beyond MEFv2 in this way, and was originally seen as a 'feature' of VS MEF. In analyzing the failure to find a way to fix the test, I came to realize that in fact that wasn't a good feature at all and caused the graph to become muddied with parts importing other parts from inappropriate sharing boundaries. So I removed the test and adjusted another to verify that v3 now matches v2 behavior.

As part of this fix, I'm enabling the GC release test that now passes with these changes.