Fix recognition of loops involving parts with CreationPolicy.Any.
When constrained by the import site to be CreationPolicy.NonShared, these would manage to throw StackOverflowException during creation.
Now we recognize these loops and throw them out with normal NonShared parts.

Also with this change, instead of failing completely to compose without any description of the faulty parts, we actually enumerate the parts involved and simply reject those parts while salvaging the rest of the composition.