Fix ComposeEditor test.
We have hard dependencies on Dev12 because StandaloneUndo.dll (which is a checked-in DLL we got from the EditorTestApp I think) only works in the Dev12 editor. The macros in the project file tried to use Dev14 assemblies which is a broken idea. It would then break Dev12 because the Dev14 file that is copied to bin\debug is newer than the Dev12 one so it wouldn't copy the right version over.

So the fix is to make sure we always copy the right file over.