Discover MEF parts in assemblies in spite of ReflectionTypeLoadException.
Some assemblies define types that derive from types defined in other assemblies that cannot be loaded. In those instances, Assembly.GetTypes() throws ReflectionTypeLoadException. Up to this point, that has caused the entire assembly to be thrown out of the MEF catalog (since we can't enumerate its types after all).

But ReflectionTypeLoadException contains an array of the types that *could* be loaded, so we can actually recover from this and enumerate all the types that are intact in the assembly, resulting in more MEF parts being discovered (perhaps all that were intended to be found).