Add resiliency against .NET changing their non-public API.
VS MEF relies on a non-public interface to be able to have exactly equivalent behavior with .NET MEF itself since they left out making an interface public. But in the anticipated event that .NET may change/rename that interface, we don't want VS MEF to be broken. So this adds a failover path where if the interface or its property is not found at runtime, we'll use a public API approach to get effectively the same result (but with _lots_ of exceptions).

There is an active bug on the BCL team to make a proper public API that we can consume that they are likely to fix. But the fix is what renames the old interface and why we need a mitigation in place in the interim.