Remove stale code generation code
Early in VS MEF development we used code generation to translate the composition to an assembly. But from the sum of code generation, compile time and JIT time, it was too slow, despite its theoretical runtime perf superiority. So we switched to always using reflection and a runtime composition system instead. Codegen has been decaying ever since. And this removes it.

See the codegen-last-stand tag (05758104af) for when codegen last worked.