Remove dependency on VS2013 being installed to build
This was for a VS2013 editor hosting test. This test was originally added as a "milestone marker" to reach for in this MEF implementation. Now that it has been reached and gone far beyond, and we have micro tests for each bit of functionality, this massive "integration" sort of test can be removed, and the huge VS2013 dependency with it.