Fix GetInputAssemblies for LazyMetadataWrapper
We were loading assemblies that should be loaded lazily before this
change. There is no requirement to load the assembly so we should check to
see if we can lazily get the assembly before loading it. Wrote some tests
to validate this, as well as renaming things in the
CompositionCatalogTests that were misleading (things weren't actually
lazy).

Also adds a test for null export metadata to make sure we work well with
null metadata.