Merged PR 65361: Provide ability to specify extra input assemblies
Provide ability to specify extra input assemblies

This is a general solution in order to support providing additional input
assemblies for ComposablePartDefinitions. The reason we're doing this is
because we want to invalidate the cache when an attribute that defines
metadata on an exported type is changed. There is no knowledge of the
attribute by the time ComposablePartDefinition.GetInputAssemblies() is
called, so the easiest/most robust solution was to support adding these
input assemblies when the ComposablePartDefinition gets created (since we
would have access to the metadata attributes at that point).

Related work items: #379277