Remove MetadataAttribute from derived metadata class
This exposes the bug where our MEFv2 discovery module ignored export metadata that is declared by custom attributes that aren't directly affixed with [MetadataAttribute], but are by virtue of a base class.
I'm also enabling V2Compat for a test method that is too long overdue and contains relevant assertions for this change.