Switch to use eagerly initialized readonly fields
I considered making them non-readonly and putting shared initialization in a helper, but without access to MemberNotNull that pattern is painful, requiring pragma suppression or lots of !s.  I also considered adding a shared private ctor, but that gets complicated for the derived types that need to delegate to two different base ctors.  Decided to just keep it simple and duplicate the initialization where necessary.