Fix symbol archival step
Symbol archival was failing when InsertConfigValues.ps1 was running from a downloaded artifact, but it is only designed to run within the repo. It failed within the repo because VSInsertion.ps1 started returning an empty set of files due to the recent SBOM change in
4a2058e: Build SBOM manifest as part of build

The fix is to give the scripts a way to communicate that the SBOM is not relevant so that InsertConfigValues can still run when required.