Build VSMefX for .NET Framework as well
This is important to avoid assembly load failures when scanning net472-targeted MEF catalog assemblies.

The dotnet tool package we were building breaks the build if the project includes .NET Framework targets, so I've disabled the package build for now. Eventually we may have to split this tool into two distinct projects if we want `dotnet tool` to continue to be a means of acquisition for the .NET Core build of the tool. But I wonder if instead we ought to have a single tool with a common front-end and then a separate process backend that can run on either .NET Core or .NET Framework as required to load the assemblies given in the catalog.