Avoid func/capture allocations for non-lazy items
Non-lazy imports were paying for capture and func creation, despite the func immediately being invoked immediately after it was created. Instead take a different path for the lazy versus non-lazy case so that only Lazy<T> exports pay for it.

This was allocating about 100 MB opening and closing Orchard Core, but the end result will depend on how many lazy versus non-lazy items are in the composition graph.