Partially revert earlier commit for always using ReadImmutableArray
It's better to let ReadList continue to do it's own thing and have ReadImmutableArray implemented separately.

One issue that I didn't notice in the profile after the earlier change was that ImmutableArray.Builder is a class, and for such small arrays, ended up allocating about as much as we were saving from not allocating duplicate arrays.

This changes the immutable array path to not use the builder the far majority of the time. In the case where the array length is > 4, we create an IEnumerable and call CreateRange using it.

Note that caching the builder didn't work well as it does for metadataBuilder as this method may be invoked recursively.