Remove Enumerable.Range
This was ugly for a couple reasons. Not only would it use a yield return state machine, but then it would potentially resize the array multiple times on growth.

Instead, go back to using ImmutableArray builders, with the caveat that we will attempt to limit the number of builders created/used.