Removed unused variables and updated access modifiers
In `AssembliesLazyLoadedTests.cs` and `CacheResiliencyTests.cs`, removed lines of code that were creating unused variables. This should not affect the functionality. Also removed a line of code in `CacheResiliencyTests.cs` that was assigning a value to `metadataToken` variable, which might affect the functionality if used later in the code. In `ExportMetadataTests.cs`, changed the access modifier of `MetadataEnumNonPublic` enum from `public` to `internal`, restricting its visibility to the current assembly. Also, removed comment lines in `ExportMetadataTests.cs`, `TestUtilities.cs`, and `Mef3DiscoveryTestCaseRunner.cs`.