Refactor variable names and update serialization methods
Renamed `arrayElementCount` to `expectedArrayElementCount` in `BaseMessagePackFormatter` and its derived classes for better readability. The `ArrayElementCount` property has also been renamed to `ExpectedArrayElementCount` and is now used in `CheckArrayHeaderCount` and `WriteArrayHeader` methods. Improved readability of array passed to `CompositeResolver.Create` in `MessagePackSerializerContext`. Updated `DeserializeData` and `Serialize` methods in `TypeRefObjectFormatter` to include serialization and deserialization of `StrongAssemblyIdentity` object and `IsShallow` property.