Update `SignFilesForPackage` target and add signing rules
Updated the `SignFilesForPackage` target in `Microsoft.VisualStudio.Composition.AppHost.targets` to execute when `MicroBuild_SigningEnabled` is `true`. Added an `ItemGroup` to the target with a new `ToolPackInputsToSign` item, which includes specific files from the `_PackageFiles` item. These files are now signed with a third-party SHA-2 Authenticode signature and are not strong-named.