Updated Formatter classes to implement IMessagePackFormatter<T>
Several classes in the `Microsoft.VisualStudio.Composition.Formatter` namespace have been refactored to implement the `IMessagePackFormatter<T>` interface instead of inheriting from the `BaseMessagePackFormatter<T>` class. The `DeserializeData` and `SerializeData` methods have been renamed to `Deserialize` and `Serialize` respectively. These methods have been updated to handle null values, check the count of elements in the array, and ensure the `reader.Depth` is decremented even if an exception is thrown. The `expectedArrayElementCount` parameter has been removed from the constructors as it is no longer needed.