`Refactor and optimize MetadataDictionaryFormatter.cs`
Refactored `MetadataDictionaryFormatter.cs` to simplify and optimize the handling of different types. The `int` type check and write process has been streamlined for better readability. The `char` and `CreationPolicy` values are now converted to `UInt16` and `UInt8` respectively before being written, possibly due to changes in the underlying system. The deserialization of `objectType` has been modified to directly use the `ReadByte()` method from `messagePackReader`, potentially simplifying the deserialization process or adapting to changes in `objectType` storage.