Updated serialization methods for MethodRef and PropertyRef classes
Refactored the serialization and deserialization process for `MethodRef` and `PropertyRef` classes. Removed the `MethodRefFormatter` class and switched to using the default MessagePack serialization and deserialization methods by adding the `[MessagePackObject]` attribute to `MethodRef` and `PropertyRef` classes. Added a new constructor to `MethodRef` with the `[SerializationConstructor]` attribute for deserialization. Updated `MethodRef` to ignore `MethodBase` and `MethodBaseNoResolve` properties and added the `Key` attribute to `Name`, `ParameterTypes`, and `GenericMethodArguments` properties for serialization order. Updated `ComposablePartDefinitionTests` to use the new `MethodRef` constructor.