Enhanced formatter resolver flexibility in serialization
The `MessagePackSerializerContext` constructor and the `GetIFormatterResolver` method have been updated to support a more flexible specification of the formatter resolver. This is achieved by allowing an `IFormatterResolver` instance to be passed as a parameter, enabling the use of custom formatter resolvers. Additionally, the internal usage of the resolver in the `GetIFormatterResolver` method has been modified to utilize the passed-in `formatterResolver` instead of the default `StandardResolverAllowPrivate.Instance`, further enhancing the configurability of the serialization process.