Add support for nested arrays
For this test, we use a metadata view member typed as a nested array to demonstrate the need for this support. I suspect that it could also manifest as an imported/exported value typed with a nested array.
In the metadata view case, this is useful when a single metadatum has a `T[]` value, and the metadata is provided by an `AllowMultiple=true` attribute, such that all metadata gets an array added to it.