Fix dependency on non-existing package
The recent introduction of a code-fixes package left the MS.VS.Composition package with a dependency on `Microsoft.VisualStudio.Composition.Analyzers.Only`, which is a package that is never built. Instead, the `Microsoft.VisualStudio.Composition.Analyzers` package now builds in the CodeFixes project. So we need to update the P2P so that the MEF package (still) depends on the package with ID `Microsoft.VisualStudio.Composition.Analyzers`.