Fix VSMEF007
The previous code was treating unattributed properties on types as imports. It was re-using logic that applies to importing constructor parameters, which are implicitly imports even when unattributed.

To fix this, we instruct the method that determines the contract as to whether implicit imports are relevant to the caller or not.