VSMEF007 considers both contract type and name
MEF uses both the contract type and name to identify a part. The prior code was essentially coalescing these into a single value when one was null. This meant that if the contract name matched a type name, then you could get a false positive for duplcate imports, where in fact the types differed.

This was observed in the dotnet/project-system repo.