Extend VSMEF007 for Lazy and ExportFactory
VSMEF007 aims to detect duplicate imports in MEF parts.

It previously did not handle `Lazy<T>` imports, and so would consider imports of both `Foo` and `Lazy<Foo>` as different. With this change, they are treated as duplicates.

This change also causes imports of `Foo` and `ExportFactory<Foo>` to be considered as duplicates. There are conceievably valid scenarios for this, though they seem unlikely and are suppressable.